trading-automation/requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
openpyxl>=3.0.0
xlrd>=2.0.1
matplotlib>=3.5.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
google-auth>=2.0.0